example(get_started): add simple hello world example
