+++ examples/get_started/lv_example_get_started.h
@@ -28,6 +28,7 @@ extern "C" {
 void lv_example_get_started_1(void);
 void lv_example_get_started_2(void);
 void lv_example_get_started_3(void);
+void lv_example_get_started_4(void);
 
 /**********************
  *      MACROS
